% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{community_gpkg_extensions}
\alias{community_gpkg_extensions}
\title{Community GeoPackage Extensions}
\format{
A data frame with 15 rows and 5 variables:
\describe{
\item{\code{pos}}{Position/order}
\item{\code{extension}}{Extension title}
\item{\code{description}}{Extension description}
\item{\code{url}}{Detail URL}
\item{\code{reference}}{List column with reference tables for
gpkg_extensions and other extension-specific tables.}
}
}
\usage{
community_gpkg_extensions
}
\description{
Reference table of community GeoPackage extensions from
\url{http://www.geopackage.org/extensions.html}. Note that the information in the
reference column is an incomplete representation of the requirements for
extension implementation. This data is provided to support a (yet to be
developed) functions that allow the addition of extensions to exported
GeoPackage files.
}
\keyword{datasets}
