% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connect_gpkg}
\alias{connect_gpkg}
\title{Connect to GeoPackage data source}
\usage{
connect_gpkg(dsn = NULL, con = NULL, call = .envir, .envir = parent.frame())
}
\arguments{
\item{dsn}{Path or url for GeoPackage file. Optional if con is provided.}

\item{con}{Connection from \code{\link[DBI:dbConnect]{DBI::dbConnect()}}. Optional if dsn is provided.}

\item{call}{The execution environment of a currently running
function, e.g. \code{call = caller_env()}. The corresponding function
call is retrieved and mentioned in error messages as the source
of the error.

You only need to supply \code{call} when throwing a condition from a
helper function which wouldn't be relevant to mention in the
message.

Can also be \code{NULL} or a \link[rlang:topic-defuse]{defused function call} to
respectively not display any call or hard-code a code to display.

For more information about error calls, see \ifelse{html}{\link[rlang:topic-error-call]{Including function calls in error messages}}{\link[rlang:topic-error-call]{Including function calls in error messages}}.}

\item{.envir}{Environment to evaluate the glue expressions in.}
}
\description{
Connect to GeoPackage data source
}
