% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg_extension.R
\name{extension_table_names}
\alias{extension_table_names}
\title{Get extension table names from the reference table of registered or community
GeoPackage extensions}
\usage{
extension_table_names(
  extension,
  community = FALSE,
  exclude = "gpkg_extensions"
)
}
\description{
Get extension table names from the reference table of registered or community
GeoPackage extensions
}
\keyword{internal}
