% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg_extension.R
\name{extension_tables}
\alias{extension_tables}
\title{Get reference tables from gpkg_extensions or community_gpkg_extensions}
\usage{
extension_tables(extension, community = FALSE)
}
\description{
Get reference tables from gpkg_extensions or community_gpkg_extensions
}
\keyword{internal}
