% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopackage.r
\name{geopackage}
\alias{geopackage}
\title{Read GeoPackage}
\usage{
geopackage(x, y, verbose = TRUE, ...)
}
\arguments{
\item{x}{geopackage file name}

\item{y}{layer name, if not specified the first layer found will be read}

\item{verbose}{report on layers available when not specified, defaults to TRUE}

\item{...}{arguments passed to \code{\link[rgdal]{readOGR}}}
}
\value{
Spatial object as per \code{\link[rgdal]{readOGR}}
}
\description{
Read Spatial data from GeoPackage format, via \code{\link[rgdal]{readOGR}}.
}
\examples{
f <- system.file("extdata", "file_poly.gpkg", package= "rGeoPackage")
x <- geopackage(f)
}

