% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gpkg_extensions}
\alias{gpkg_extensions}
\title{Registered GeoPackage Extensions}
\format{
A data frame with 9 rows and 4 variables:
\describe{
\item{\code{pos}}{Position/order}
\item{\code{extension}}{Extension title}
\item{\code{url}}{Detail URL}
\item{\code{reference}}{List column with reference tables for
gpkg_extensions and other extension-specific tables.}
}
}
\usage{
gpkg_extensions
}
\description{
Reference table of registered GeoPackage extensions from
\url{http://www.geopackage.org/guidance/getting-started.html#registered-extensions}.
Note that the information in the reference column is an incomplete
representation of the requirements for extension implementation. This data is
provided to support a (yet to be developed) functions that allow the addition
of extensions to exported GeoPackage files.
}
\keyword{datasets}
