% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gpkg_metadata.R
\name{read_gpkg_metadata}
\alias{read_gpkg_metadata}
\alias{list_gpkg_metadata}
\title{Read or list GeoPackage metadata with sf}
\usage{
read_gpkg_metadata(dsn, scope = "dataset", as_xml = TRUE, ...)

list_gpkg_metadata(dsn, ...)
}
\arguments{
\item{dsn}{Data source name. Must be a GeoPackage file.}

\item{scope}{Metadata scope.}

\item{as_xml}{If \code{TRUE} (default), parse the metadata to xml with
\code{\link[xml2:read_xml]{xml2::read_xml()}}. If the specified scope matches more than one metadata
entry, this function returns a list of xml objects.}

\item{...}{Additional parameters passed to \code{\link[sf:st_read]{sf::read_sf()}}. Not required.}
}
\description{
Read or list GeoPackage metadata with sf
}
