% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gpkg_table.R
\name{read_gpkg_table}
\alias{read_gpkg_table}
\title{Read a GeoPackage table with sf}
\usage{
read_gpkg_table(
  dsn,
  name,
  query = NULL,
  select = NULL,
  where = NULL,
  call = parent.frame(),
  ...
)
}
\arguments{
\item{dsn}{Data source name. Must be a GeoPackage file.}

\item{name}{Name of a GeoPackage table to read.}

\item{query}{If \code{NULL}, query defaults to "SELECT {select} FROM {name}". If
query is provided, select and name are ignored.}

\item{where}{If not \code{NULL}, where is appended to the end of query as "{query}
WHERE {where}".}

\item{call}{Passed to \code{\link[rlang:check_required]{rlang::check_required()}}}

\item{...}{Additional parameters passed to \code{\link[sf:st_read]{sf::read_sf()}}. Not required.}
}
\description{
Read a GeoPackage table with sf
}
