% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeopackage-package.R
\docType{package}
\name{rgeopackage-package}
\alias{rgeopackage}
\alias{rgeopackage-package}
\title{rgeopackage: Helper Tools in Creating or Handling GeoPackage Files}
\description{
The aim of this R package is to provide helper tools in creating or handling GeoPackage files, in order to complement other R spatial packages or GDAL. The package has no dependencies on other spatial packages and is not tied to any particular package by design.
}
\author{
\strong{Maintainer}: Floris Vanderhaeghe \email{floris.vanderhaeghe@inbo.be} (\href{https://orcid.org/0000-0002-6378-6229}{ORCID})

Authors:
\itemize{
  \item Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [contributor]
}

Other contributors:
\itemize{
  \item Research Institute for Nature and Forest \email{info@inbo.be} [copyright holder]
}

}
\keyword{internal}
