% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gpkg.R
\name{write_gpkg_table}
\alias{write_gpkg_table}
\alias{update_gpkg_table}
\alias{append_gpkg_table}
\alias{create_gpkg_table}
\title{Create, append, or update a GeoPackage table}
\usage{
write_gpkg_table(
  dsn = NULL,
  con = NULL,
  table_name = NULL,
  table_data,
  overwrite = FALSE,
  append = FALSE,
  header = TRUE,
  temporary = FALSE,
  quiet = FALSE,
  call = parent.frame(),
  ...
)

update_gpkg_table(
  con,
  table_name = NULL,
  statement = NULL,
  message = NULL,
  quiet = FALSE,
  .envir = parent.frame()
)

append_gpkg_table(
  dsn = NULL,
  con = NULL,
  table_name,
  table_data,
  call = parent.frame(),
  ...
)

create_gpkg_table(
  dsn = NULL,
  con = NULL,
  table_name,
  fields,
  quiet = FALSE,
  temporary = FALSE,
  ...
)
}
\arguments{
\item{dsn}{Path or url for GeoPackage file. Optional if con is provided.}

\item{con}{Connection from \code{\link[DBI:dbConnect]{DBI::dbConnect()}}. Optional if dsn is provided.}

\item{table_name}{Name of a GeoPackage table to read}

\item{temporary}{If \code{TRUE}, will generate a temporary table statement.}

\item{quiet}{If \code{TRUE}, do not show informational messages. Defaults to
\code{FALSE}.}

\item{call}{The execution environment of a currently running
function, e.g. \code{call = caller_env()}. The corresponding function
call is retrieved and mentioned in error messages as the source
of the error.

You only need to supply \code{call} when throwing a condition from a
helper function which wouldn't be relevant to mention in the
message.

Can also be \code{NULL} or a \link[rlang:topic-defuse]{defused function call} to
respectively not display any call or hard-code a code to display.

For more information about error calls, see \ifelse{html}{\link[rlang:topic-error-call]{Including function calls in error messages}}{\link[rlang:topic-error-call]{Including function calls in error messages}}.}

\item{...}{
  Arguments passed on to \code{\link[DBI:dbWriteTable]{DBI::dbWriteTable}}
  \describe{
    \item{\code{conn}}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[DBI:dbConnect]{dbConnect()}}.}
    \item{\code{name}}{The table name, passed on to \code{\link[DBI:dbQuoteIdentifier]{dbQuoteIdentifier()}}. Options are:
\itemize{
\item a character string with the unquoted DBMS table name,
e.g. \code{"table_name"},
\item a call to \code{\link[DBI:Id]{Id()}} with components to the fully qualified table name,
e.g. \code{Id(schema = "my_schema", table = "table_name")}
\item a call to \code{\link[DBI:SQL]{SQL()}} with the quoted and fully qualified table name
given verbatim, e.g. \code{SQL('"my_schema"."table_name"')}
}}
    \item{\code{value}}{a \link{data.frame} (or coercible to data.frame).}
  }}

\item{fields}{Either a character vector or a data frame.

A named character vector: Names are column names, values are types.
Names are escaped with \code{\link[DBI:dbQuoteIdentifier]{dbQuoteIdentifier()}}.
Field types are unescaped.

A data frame: field types are generated using
\code{\link[DBI:dbDataType]{dbDataType()}}.}
}
\description{
Create, append, or update a GeoPackage table
}
