% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg_contents.R
\name{read_gpkg_contents}
\alias{read_gpkg_contents}
\alias{update_gpkg_contents}
\title{Read and update GeoPackage contents table}
\usage{
read_gpkg_contents(dsn)

update_gpkg_contents(
  dsn,
  identifier = NULL,
  description = NULL,
  title = identifier,
  timestamp = Sys.time(),
  quiet = FALSE
)
}
\description{
Read and update GeoPackage contents table
}
